osd,common: avoid casting useless casts
no need to cast a variable of type `Foo` to `Foo`, for example, 2.000 is
of type `double`, so (double)2.000 does not make sense.

+++ src/include/utime.h
@@ -79,7 +79,7 @@ public:
   }
   void set_from_double(double d) { 
     tv.tv_sec = (__u32)trunc(d);
-    tv.tv_nsec = (__u32)((d - (double)tv.tv_sec) * (double)1000000000.0);
+    tv.tv_nsec = (__u32)((d - (double)tv.tv_sec) * 1000000000.0);
   }
 
   real_time to_real_time() const {
@@ -404,7 +404,7 @@ inline utime_t& operator+=(utime_t& l, const utime_t& r) {
 }
 inline utime_t& operator+=(utime_t& l, double f) {
   double fs = trunc(f);
-  double ns = (f - fs) * (double)1000000000.0;
+  double ns = (f - fs) * 1000000000.0;
   l.sec_ref() += (long)fs;
   l.nsec_ref() += (long)ns;
   l.normalize();
@@ -427,7 +427,7 @@ inline utime_t& operator-=(utime_t& l, const utime_t& r) {
 }
 inline utime_t& operator-=(utime_t& l, double f) {
   double fs = trunc(f);
-  double ns = (f - fs) * (double)1000000000.0;
+  double ns = (f - fs) * 1000000000.0;
   l.sec_ref() -= (long)fs;
   long nsl = (long)ns;
   if (nsl) {
